Isolate models in a separate file
