+++ miniscope_io/cli/update.py
@@ -4,7 +4,8 @@ CLI for updating device over IR or UART.
 
 import click
 
-from miniscope_io.device_update import DevUpdate, DeviceCommand
+from miniscope_io.device_update import DeviceCommand, DevUpdate
+
 
 @click.command()
 @click.option(
@@ -39,7 +40,7 @@ from miniscope_io.device_update import DevUpdate, DeviceCommand
     "--restart",
     is_flag=True,
     type=bool,
-    help="Restart the device. Cannot be used with --target or --value."
+    help="Restart the device. Cannot be used with --target or --value.",
 )
 def update(port: str, target: str, value: int, device_id: int, restart: bool) -> None:
     """
@@ -49,17 +50,13 @@ def update(port: str, target: str, value: int, device_id: int, restart: bool) ->
     # Check mutual exclusivity
     if (target and not value) or (value and not target):
         raise click.UsageError("Both --target and --value are required if one is specified.")
-    
+
     if (target or value) and restart:
         raise click.UsageError("Options --target/--value and --restart cannot be used together.")
 
     if target and value:
         DevUpdate(port=port, target=target, value=value, device_id=device_id)
     elif restart:
-        DevUpdate(port=port,
-                  target='DEVICE',
-                  value=DeviceCommand.RESTART,
-                  device_id=device_id
-                  )
+        DevUpdate(port=port, target="DEVICE", value=DeviceCommand.RESTART, device_id=device_id)
     else:
         raise click.UsageError("Either --target with --value or --restart must be specified.")
